Crate cwe_checker

This crate defines the command line interface for the cwe_checker. General documentation about the cwe_checker is contained in the cwe_checker_lib crate.

Structs

CmdlineArgs
Find vulnerable patterns in binary executables

Functions

check_file_existence
Check the existence of a file

filter_modules_for_partial_run
Only keep the modules specified by the --partial parameter in the modules list. The parameter is a comma-separated list of module names, e.g. ‘CWE332,CWE476,CWE782’.

get_project_from_ghidra
Execute the p_code_extractor plugin in ghidra and parse its output into the Project data structure.

main

run_with_ghidra
Run the cwe_checker with Ghidra as its backend.