Struct cwe_checker::CmdlineArgs

pub(crate) struct CmdlineArgs {
    binary: Option<String>,
    config: Option<String>,
    out: Option<String>,
    partial: Option<String>,
    json: bool,
    quiet: bool,
    module_versions: bool,
    debug: bool,
}

Find vulnerable patterns in binary executables

Fields
binary: Option<String>

The path to the binary.

config: Option<String>

Path to a custom configuration file to use instead of the standard one.

out: Option<String>

Write the results to a file instead of stdout. This only affects CWE warnings. Log messages are still printed to stdout.

partial: Option<String>

Specify a specific set of checks to be run as a comma separated list, e.g. ‘CWE332,CWE476,CWE782’.

Use the “–module-versions” command line option to get a list of all valid check names.

json: bool

Generate JSON output.

quiet: bool

Do not print log messages. This prevents polluting stdout for json output.

module_versions: bool

Prints out the version numbers of all known modules.

debug: bool

Output for debugging purposes. The current behavior of this flag is unstable and subject to change.